Working out a few niggles in the Spectrum View last night opened a minor rats' nest of small UI bugs.  I apologize for this all being monolithic, but they were all interelated.
1) Spectrum view was visually off-center with repect to the waveform;
to see this bug in an old version generate silence, and draw one
sample to nonzero.  This patch centers the FFT response, also
offsetting by 1/2 bin to reflect the overall metaphor in Audacity that
sample points are a regular sequence from 0, 1, 2... rather than .5,
1.5, 2.5...

2) Added block sizes down to 8 samples wide; yes, I actually use this :-)

3) Clicking 'OK' in the prefs panel would crash if there was no valid
audio device configurable; the bug was due to addressing a zero-element
vector of strings.

4) The Spectrum View cache had a buffer overflow bug when looking at
the end of a sequence, could cuase a crash.

5) the 'draw tool' did not mark the WaveTrack data changed when
altering samples due to calling 'Sequence::Set' directly rather than
the wrapper 'WaveTrack::Set'.  Thi swas probably due to WaveTrack::Set
internally re-applying the track start time offset when the draw code
was already taking that into account.  However, only the draw tool
ever called WaveTrack::Set anyway, so I modified the behavior to be
more useful, altered the draw tool to call WaveTrack::Set() and now
cache dirtying works properly.

6) related to 5), the 'smooth tool' *was* calling WaveTrack::Set, but
as both the smooth tool and WaveTrack::Set were adding the start track
time offset, all smooth operations were happening in the wrong place.
Changing the ::Set behavior to not apply time offset in 5) also fixes
6).

7) Changing spectral preferences did not trigger a redraw upon
application of new prefs; added a SpecCache state var to look for change
in window size and desired upper frequency bound.

8) Spectral view (and subvariants) generated FPU exceptions (NaN and
-Inf) when looking at silence due to dB(0.) being -Inf.  Added dB
approximation code from Postfish that is both guarded against -Inf and
faster, but requires an IEEE754 compliant CPU.  This also corrects the
visible black lines that result.

9) Clarified "draw only works on waveforms" message to "the draw tool
currently only works on linear view waveform tracks"

10) the beginnings and ends of WaveTracks (in linear, db and spectral
view) generally overran the boundaries of the track.  In order to
present a consistent track metaphor, waveform/spectral display and
selection now begin at the first point in a sequence and end at the
last point, period.

11) correct spectral and waveform view to present consistent selection
boundary display across tracks (cosmetic issue only).

12) Correct selection dragging in non-snap-to-mode to a more intuitive
behavior (select all points dragged over) and change display so that
when zoomed to individual points, selection is displayed beginning and
ending between points in the selection so that it is unambiguous what
points are selected.

13) As above for snap-to modes, but these introduce a problem; because
snap-to defines selection beginning/end to occur exactly at a sample
point, it is ambiguous if the beginning and end samples are included.
In fact, in different modes, the last point was sometimes included,
sometimes not.  Also, because samples are at the beginning of a sample
period, not centered, the selection of a subsequence must extend
forward by one phantom point, otherwise it's impossible to select only
one point in a track (or select the last point in any track).  End
compromise: Leave actual snap-to select behavior and ruler display
unchanged (as it might be slightly confusing at first, but works and
allows all use cases) and alter track display behavior to show,
unambiguously, which points are selected as in 12) above.