Reworked Track, TrackList, and TrackListIterator classes so they take advantage of the double linked list instead of constantly searching as if it were singly linked.  Major speed improvements.
Added TrackListOfKindIterator to alloc scanning tracklists for a
specific track type.

Removed GetTeamed() and SetTeamed() since the teaming was not being
updated in all cases.  This was causing a least one bug in TrackPanel
navigation.

Added a GetLink() method to Track to make it easier to get at the linked
track and to detect linked tracks since GetTeamed() was removed.  Took
advantage of it in a few, but not all, locations.

Simplified AudacityProject::OnDuplicate() to remove dependence on copy
constructor of TracklistIterator (since it no longer exists.

Changed TrackListIterator::RemoveCurrent() so it not deletes the track
if requested.  Otherwise it sets the node backpointer to NULL.

Removed double duplication of the tracklist in AudacityProject methods
InitialState(), PushState(), and ModifyState.

Any function where I made "significant" changes, I reformatted to a
single coding style.  I know this is unnecessary, but some functions
seemed to have like 4 or more different styles.

Modified TrackPanel::HandleRearrange() to allow moving tracks when top
or bottom of window is reached.  Prior to this, you had to stop
dragging, scroll the window manually, and continue dragging.

Modified TrackPanel functions HandleResizeClick() and HandleResizeDrag()
to not take tracks out of minimize state until an actual drag occurs.
And, when it does, to change the state of both tracks.

Reworked TrackPanel functions RedrawEverythingElse(), DrawOutside(), and
HighlightFocusedTrack() to only draw the visible tracks.  Major speed
improvement.

Made TrackPanel::EnsureVisible() only scroll down as much as is needed to
ensure that a track off the bottom of the panel is displayed entirely.
Before it would be scrolled to the top of the window.

Corrected deletion of tracks on the undo stack.

Changed Effect to no longer make a copy of the project's tracklist as the
effects now take advantage of the new TrackListOfKindIterator.  This also
simplifies the track copying and replacing prior/following applying the
effect.

Fixed an issue where Effect::Preview would get the wrong number of tracks.

Fixed several null pointer dereferences and memory leaks.